Implement thread IDs based on @cehteh's work
This adds features named "threads" and "nightly".

Enabling "threads" and calling `.with_threads(true)` on the logger
before init will show thread names in the output. If a thread is not
named, '?' will be shown.

Enabling "threads" and "nightly" will show the numeric thread ID if
a thread does not have a name.

This is based on https://github.com/borntyping/rust-simple_logger/pull/39