Make UTC the default timestamps option, and change the output format to include the timezone
This works around the issue of potentially silently changing the
timezone used in logfiles (and not leaving the user any way to notice
the change) by changing the timezone format used from one that is
timezone agnostic to one that includes the timezone.

Since it's alreday changing from one that perfectly matches Supervisord,
this uses the well known RFC 3339 format that implements ISO 8601 (and
happens to currently be the only format provided by the `time` crate
that I don't have to implement myself).